refactor(www): Split up gatsby-node (#21217)
* Split onCreateNode

* Split createPages

* Group contributor pages with docs

* Improve graphql queries

* Use async/await instead of promises

* Separate features

* Improve code quality
