+++ www/gatsby-node.js
@@ -1,70 +1,21 @@
-const _ = require(`lodash`)
 const Promise = require(`bluebird`)
 const fetch = require(`node-fetch`)
-const path = require(`path`)
 const fs = require(`fs-extra`)
-const { slash } = require(`gatsby-core-utils`)
-const slugify = require(`slugify`)
-const url = require(`url`)
-const getpkgjson = require(`get-package-json-from-github`)
-const parseGHUrl = require(`parse-github-url`)
-const { GraphQLClient } = require(`@jamo/graphql-request`)
-const moment = require(`moment`)
 const startersRedirects = require(`./starter-redirects.json`)
-const langs = require("./i18n.json")
-const {
-  generateComparisonPageSet,
-} = require(`./src/utils/generate-comparison-page-set.js`)
-const { getPrevAndNext } = require(`./src/utils/get-prev-and-next.js`)
-const { localizedPath } = require(`./src/utils/i18n.js`)
 const yaml = require(`js-yaml`)
-const ecosystemFeaturedItems = yaml.load(
-  fs.readFileSync(`./src/data/ecosystem/featured-items.yaml`)
-)
 const redirects = yaml.load(fs.readFileSync(`./redirects.yaml`))
 
-if (
-  process.env.gatsby_executing_command === `build` &&
-  !process.env.GITHUB_API_TOKEN
-) {
-  throw new Error(
-    `A GitHub token is required to build the site. Check the README.`
-  )
-}
+const docs = require(`./src/utils/node/docs.js`)
+const showcase = require(`./src/utils/node/showcase.js`)
+const starters = require(`./src/utils/node/starters.js`)
+const creators = require(`./src/utils/node/creators.js`)
+const packages = require(`./src/utils/node/packages.js`)
+const features = require(`./src/utils/node/features.js`)
+const sections = [docs, showcase, starters, creators, packages, features]
 
-// used to gather repo data on starters
-const githubApiClient = process.env.GITHUB_API_TOKEN
-  ? new GraphQLClient(`https://api.github.com/graphql`, {
-      headers: {
-        authorization: `Bearer ${process.env.GITHUB_API_TOKEN}`,
-      },
-    })
-  : null
-
-const localPackages = `../packages`
-const localPackagesArr = []
-fs.readdirSync(localPackages).forEach(file => {
-  localPackagesArr.push(file)
-})
-// convert a string like `/some/long/path/name-of-docs/` to `name-of-docs`
-const slugToAnchor = slug =>
-  slug
-    .split(`/`) // split on dir separators
-    .filter(item => item !== ``) // remove empty values
-    .pop() // take last item
-
-const docSlugFromPath = parsedFilePath => {
-  if (parsedFilePath.name !== `index` && parsedFilePath.dir !== ``) {
-    return `/${parsedFilePath.dir}/${parsedFilePath.name}/`
-  } else if (parsedFilePath.dir === ``) {
-    return `/${parsedFilePath.name}/`
-  } else {
-    return `/${parsedFilePath.dir}/`
-  }
-}
-
-exports.createPages = ({ graphql, actions, reporter }) => {
-  const { createPage, createRedirect } = actions
+exports.createPages = async helpers => {
+  const { actions } = helpers
+  const { createRedirect } = actions
 
   redirects.forEach(redirect => {
     createRedirect({ isPermanent: true, ...redirect })
@@ -78,600 +29,12 @@ exports.createPages = ({ graphql, actions, reporter }) => {
     })
   })
 
-  return new Promise((resolve, reject) => {
-    const docsTemplate = path.resolve(`src/templates/template-docs-markdown.js`)
-    const apiTemplate = path.resolve(`src/templates/template-api-markdown.js`)
-    const blogPostTemplate = path.resolve(`src/templates/template-blog-post.js`)
-    const blogListTemplate = path.resolve(`src/templates/template-blog-list.js`)
-    const tagTemplate = path.resolve(`src/templates/tags.js`)
-    const contributorPageTemplate = path.resolve(
-      `src/templates/template-contributor-page.js`
-    )
-    const localPackageTemplate = path.resolve(
-      `src/templates/template-docs-local-packages.js`
-    )
-    const remotePackageTemplate = path.resolve(
-      `src/templates/template-docs-remote-packages.js`
-    )
-    const showcaseTemplate = path.resolve(
-      `src/templates/template-showcase-details.js`
-    )
-    const creatorPageTemplate = path.resolve(
-      `src/templates/template-creator-details.js`
-    )
-    const featureComparisonPageTemplate = path.resolve(
-      `src/templates/template-feature-comparison.js`
-    )
-
-    // Query for markdown nodes to use in creating pages.
-    graphql(`
-      query {
-        allMdx(
-          sort: { order: DESC, fields: [frontmatter___date, fields___slug] }
-          limit: 10000
-          filter: { fileAbsolutePath: { ne: null } }
-        ) {
-          edges {
-            node {
-              fields {
-                slug
-                locale
-                package
-                released
-              }
-              frontmatter {
-                title
-                draft
-                canonicalLink
-                publishedAt
-                issue
-                tags
-                jsdoc
-                apiCalls
-              }
-            }
-          }
-        }
-        allAuthorYaml {
-          edges {
-            node {
-              fields {
-                slug
-              }
-            }
-          }
-        }
-        allCreatorsYaml {
-          edges {
-            node {
-              name
-              fields {
-                slug
-              }
-            }
-          }
-        }
-        allSitesYaml(filter: { main_url: { ne: null } }) {
-          edges {
-            node {
-              main_url
-              fields {
-                slug
-                hasScreenshot
-              }
-            }
-          }
-        }
-        allStartersYaml {
-          edges {
-            node {
-              id
-              fields {
-                starterShowcase {
-                  slug
-                  stub
-                }
-                hasScreenshot
-              }
-              url
-              repo
-            }
-          }
-        }
-        allNpmPackage {
-          edges {
-            node {
-              id
-              title
-              slug
-            }
-          }
-        }
-      }
-    `).then(result => {
-      if (result.errors) {
-        return reject(result.errors)
-      }
-
-      const blogPosts = _.filter(result.data.allMdx.edges, edge => {
-        const slug = _.get(edge, `node.fields.slug`)
-        const draft = _.get(edge, `node.frontmatter.draft`)
-        if (!slug) return undefined
-
-        if (_.includes(slug, `/blog/`) && !draft) {
-          return edge
-        }
-
-        return undefined
-      })
-
-      const releasedBlogPosts = blogPosts.filter(post =>
-        _.get(post, `node.fields.released`)
-      )
-
-      // Create blog-list pages.
-      const postsPerPage = 8
-      const numPages = Math.ceil(releasedBlogPosts.length / postsPerPage)
-
-      Array.from({
-        length: numPages,
-      }).forEach((_, i) => {
-        createPage({
-          path: i === 0 ? `/blog` : `/blog/page/${i + 1}`,
-          component: slash(blogListTemplate),
-          context: {
-            limit: postsPerPage,
-            skip: i * postsPerPage,
-            numPages,
-            currentPage: i + 1,
-          },
-        })
-      })
-
-      // Create blog-post pages.
-      blogPosts.forEach((edge, index) => {
-        let next = index === 0 ? null : blogPosts[index - 1].node
-        if (next && !_.get(next, `fields.released`)) next = null
-
-        const prev =
-          index === blogPosts.length - 1 ? null : blogPosts[index + 1].node
-
-        createPage({
-          path: `${edge.node.fields.slug}`, // required
-          component: slash(blogPostTemplate),
-          context: {
-            slug: edge.node.fields.slug,
-            prev,
-            next,
-          },
-        })
-      })
-
-      const makeSlugTag = tag => _.kebabCase(tag.toLowerCase())
-
-      // Collect all tags and group them by their kebab-case so that
-      // hyphenated and spaced tags are treated the same. e.g
-      // `case-study` -> [`case-study`, `case study`]. The hyphenated
-      // version will be used for the slug, and the spaced version
-      // will be used for human readability (see templates/tags)
-      const tagGroups = _(releasedBlogPosts)
-        .map(post => _.get(post, `node.frontmatter.tags`))
-        .filter()
-        .flatten()
-        .uniq()
-        .groupBy(makeSlugTag)
-
-      tagGroups.forEach((tags, tagSlug) => {
-        createPage({
-          path: `/blog/tags/${tagSlug}/`,
-          component: tagTemplate,
-          context: {
-            tags,
-          },
-        })
-      })
-
-      // Create starter pages.
-      const starters = _.filter(result.data.allStartersYaml.edges, edge => {
-        const slug = _.get(edge, `node.fields.starterShowcase.slug`)
-        if (!slug) {
-          return null
-        } else if (!_.get(edge, `node.fields.hasScreenshot`)) {
-          reporter.warn(
-            `Starter showcase entry "${edge.node.repo}" seems offline. Skipping.`
-          )
-          return null
-        } else {
-          return edge
-        }
-      })
-
-      const starterTemplate = path.resolve(
-        `src/templates/template-starter-page.js`
-      )
-
-      starters.forEach((edge, index) => {
-        createPage({
-          path: `/starters${edge.node.fields.starterShowcase.slug}`,
-          component: slash(starterTemplate),
-          context: {
-            slug: edge.node.fields.starterShowcase.slug,
-            stub: edge.node.fields.starterShowcase.stub,
-          },
-        })
-      })
-
-      // Create contributor pages.
-      result.data.allAuthorYaml.edges.forEach(edge => {
-        createPage({
-          path: `${edge.node.fields.slug}`,
-          component: slash(contributorPageTemplate),
-          context: {
-            slug: edge.node.fields.slug,
-          },
-        })
-      })
-
-      result.data.allCreatorsYaml.edges.forEach(edge => {
-        if (!edge.node.fields) return
-        if (!edge.node.fields.slug) return
-        createPage({
-          path: `${edge.node.fields.slug}`,
-          component: slash(creatorPageTemplate),
-          context: {
-            slug: edge.node.fields.slug,
-            name: edge.node.name,
-          },
-        })
-      })
-
-      result.data.allSitesYaml.edges.forEach(edge => {
-        if (!edge.node.fields) return
-        if (!edge.node.fields.slug) return
-        if (!edge.node.fields.hasScreenshot) {
-          reporter.warn(
-            `Site showcase entry "${edge.node.main_url}" seems offline. Skipping.`
-          )
-          return
-        }
-        createPage({
-          path: `${edge.node.fields.slug}`,
-          component: slash(showcaseTemplate),
-          context: {
-            slug: edge.node.fields.slug,
-          },
-        })
-      })
-
-      // Create docs pages.
-      const docPages = result.data.allMdx.edges
-
-      docPages.forEach(({ node }) => {
-        const slug = _.get(node, `fields.slug`)
-        const locale = _.get(node, `fields.locale`)
-        if (!slug) return
-
-        if (!_.includes(slug, `/blog/`)) {
-          const prevAndNext = getPrevAndNext(node.fields.slug)
-
-          if (node.frontmatter.jsdoc) {
-            // API template
-            createPage({
-              path: localizedPath(locale, node.fields.slug),
-              component: slash(apiTemplate),
-              context: {
-                slug: node.fields.slug,
-                jsdoc: node.frontmatter.jsdoc,
-                apiCalls: node.frontmatter.apiCalls,
-                ...prevAndNext,
-              },
-            })
-          } else if (node.fields.package) {
-            // Local package template
-            createPage({
-              path: `${node.fields.slug}`,
-              component: slash(localPackageTemplate),
-              context: {
-                slug: node.fields.slug,
-              },
-            })
-          } else {
-            // Docs template
-            createPage({
-              path: localizedPath(locale, node.fields.slug),
-              component: slash(docsTemplate),
-              context: {
-                slug: node.fields.slug,
-                locale,
-                ...prevAndNext,
-              },
-            })
-          }
-        }
-      })
-
-      const allPackages = result.data.allNpmPackage.edges
-      // Create package readme
-      allPackages.forEach(edge => {
-        if (_.includes(localPackagesArr, edge.node.title)) {
-          createPage({
-            path: edge.node.slug,
-            component: slash(localPackageTemplate),
-            context: {
-              slug: edge.node.slug,
-              id: edge.node.id,
-            },
-          })
-        } else {
-          createPage({
-            path: edge.node.slug,
-            component: slash(remotePackageTemplate),
-            context: {
-              slug: edge.node.slug,
-              id: edge.node.id,
-            },
-          })
-        }
-      })
-
-      // Create feature comparison pages
-      const jamstackPages = generateComparisonPageSet(`jamstack`)
-      const cmsPages = generateComparisonPageSet(`cms`)
-      const comparisonPages = [...jamstackPages, ...cmsPages]
-      for (const { path, options, featureType } of comparisonPages) {
-        createPage({
-          path,
-          component: slash(featureComparisonPageTemplate),
-          context: {
-            options,
-            featureType,
-          },
-        })
-      }
-
-      return resolve()
-    })
-  })
+  await Promise.all(sections.map(section => section.createPages(helpers)))
 }
 
 // Create slugs for files, set released status for blog posts.
-exports.onCreateNode = ({ node, actions, getNode, reporter }) => {
-  const { createNodeField } = actions
-  let slug
-  let locale
-  if (node.internal.type === `File`) {
-    const parsedFilePath = path.parse(node.relativePath)
-    // TODO add locale data for non-MDX files
-    if (node.sourceInstanceName === `docs`) {
-      slug = docSlugFromPath(parsedFilePath)
-    }
-    if (slug) {
-      createNodeField({ node, name: `slug`, value: slug })
-    }
-  } else if (
-    [`MarkdownRemark`, `Mdx`].includes(node.internal.type) &&
-    getNode(node.parent).internal.type === `File`
-  ) {
-    const fileNode = getNode(node.parent)
-    const parsedFilePath = path.parse(fileNode.relativePath)
-    // Add slugs for docs pages
-    if (fileNode.sourceInstanceName === `docs`) {
-      slug = docSlugFromPath(parsedFilePath)
-      locale = "en"
-
-      // Set released status and `published at` for blog posts.
-      if (_.includes(parsedFilePath.dir, `blog`)) {
-        let released = false
-        const date = _.get(node, `frontmatter.date`)
-        if (date) {
-          released = moment.utc().isSameOrAfter(moment.utc(date))
-        }
-        createNodeField({ node, name: `released`, value: released })
-
-        const canonicalLink = _.get(node, `frontmatter.canonicalLink`)
-        const publishedAt = _.get(node, `frontmatter.publishedAt`)
-
-        createNodeField({
-          node,
-          name: `publishedAt`,
-          value: canonicalLink
-            ? publishedAt || url.parse(canonicalLink).hostname
-            : null,
-        })
-      }
-    }
-
-    for (let { code } of langs) {
-      if (fileNode.sourceInstanceName === `docs-${code}`) {
-        // have to remove the beginning "/docs" path because of the way
-        // gatsby-source-filesystem and gatsby-source-git differ
-        slug = docSlugFromPath(path.parse(fileNode.relativePath.substring(5)))
-        locale = code
-      }
-    }
-
-    // Add slugs for package READMEs.
-    if (
-      fileNode.sourceInstanceName === `packages` &&
-      parsedFilePath.name === `README`
-    ) {
-      slug = `/packages/${parsedFilePath.dir}/`
-      createNodeField({
-        node,
-        name: `title`,
-        value: parsedFilePath.dir,
-      })
-      createNodeField({ node, name: `package`, value: true })
-    }
-    if (slug) {
-      createNodeField({ node, name: `anchor`, value: slugToAnchor(slug) })
-      createNodeField({ node, name: `slug`, value: slug })
-    }
-    if (locale) {
-      createNodeField({ node, name: `locale`, value: locale })
-    }
-  } else if (node.internal.type === `AuthorYaml`) {
-    slug = `/contributors/${slugify(node.id, {
-      lower: true,
-    })}/`
-    createNodeField({ node, name: `slug`, value: slug })
-  } else if (node.internal.type === `SitesYaml` && node.main_url) {
-    const parsed = url.parse(node.main_url)
-    const cleaned = parsed.hostname + parsed.pathname
-    slug = `/showcase/${slugify(cleaned)}`
-    createNodeField({ node, name: `slug`, value: slug })
-
-    // determine if screenshot is available
-    const screenshotNode = node.children
-      .map(childID => getNode(childID))
-      .find(node => node.internal.type === `Screenshot`)
-
-    createNodeField({ node, name: `hasScreenshot`, value: !!screenshotNode })
-  } else if (node.internal.type === `NPMPackage`) {
-    if (ecosystemFeaturedItems.plugins.includes(node.name)) {
-      createNodeField({ node, name: `featured`, value: true })
-    }
-  } else if (node.internal.type === `StartersYaml` && node.repo) {
-    // To develop on the starter showcase, you'll need a GitHub
-    // personal access token. Check the `www` README for details.
-    // Default fields are to avoid graphql errors.
-    const { owner, name: repoStub } = parseGHUrl(node.repo)
-
-    // mark if it's a featured starter
-    if (ecosystemFeaturedItems.starters.includes(`/${owner}/${repoStub}/`)) {
-      createNodeField({ node, name: `featured`, value: true })
-    }
-
-    const defaultFields = {
-      slug: `/${owner}/${repoStub}/`,
-      stub: repoStub,
-      name: ``,
-      description: ``,
-      stars: 0,
-      lastUpdated: ``,
-      owner: ``,
-      githubFullName: ``,
-      gatsbyMajorVersion: [[`no data`, `0`]],
-      allDependencies: [[`no data`, `0`]],
-      gatsbyDependencies: [[`no data`, `0`]],
-      miscDependencies: [[`no data`, `0`]],
-    }
-
-    // determine if screenshot is available
-    const screenshotNode = node.children
-      .map(childID => getNode(childID))
-      .find(node => node.internal.type === `Screenshot`)
-
-    createNodeField({ node, name: `hasScreenshot`, value: !!screenshotNode })
-
-    if (!process.env.GITHUB_API_TOKEN) {
-      return createNodeField({
-        node,
-        name: `starterShowcase`,
-        value: {
-          ...defaultFields,
-        },
-      })
-    } else {
-      return Promise.all([
-        getpkgjson(node.repo),
-        githubApiClient.request(`
-            query {
-              repository(owner:"${owner}", name:"${repoStub}") {
-                name
-                stargazers {
-                  totalCount
-                }
-                createdAt
-                pushedAt
-                owner {
-                  login
-                }
-                nameWithOwner
-              }
-            }
-          `),
-      ])
-        .then(results => {
-          const [pkgjson, githubData] = results
-          const {
-            stargazers: { totalCount: stars },
-            pushedAt: lastUpdated,
-            owner: { login: owner },
-            name,
-            nameWithOwner: githubFullName,
-          } = githubData.repository
-
-          const { dependencies = [], devDependencies = [] } = pkgjson
-          const allDependencies = Object.entries(dependencies).concat(
-            Object.entries(devDependencies)
-          )
-
-          const gatsbyMajorVersion = allDependencies
-            .filter(([key, _]) => key === `gatsby`)
-            .map(version => {
-              let [gatsby, versionNum] = version
-              if (versionNum === `latest` || versionNum === `next`) {
-                return [gatsby, `2`]
-              }
-              return [gatsby, versionNum.replace(/\D/g, ``).charAt(0)]
-            })
-
-          // If a new field is added here, make sure a corresponding
-          // change is made to "defaultFields" to not break DX
-          const starterShowcaseFields = {
-            slug: `/${owner}/${repoStub}/`,
-            stub: repoStub,
-            name,
-            description: pkgjson.description,
-            stars,
-            lastUpdated,
-            owner,
-            githubFullName,
-            gatsbyMajorVersion,
-            allDependencies,
-            gatsbyDependencies: allDependencies
-              .filter(
-                ([key, _]) => ![`gatsby-cli`, `gatsby-link`].includes(key) // remove stuff everyone has
-              )
-              .filter(([key, _]) => key.includes(`gatsby`)),
-            miscDependencies: allDependencies.filter(
-              ([key, _]) => !key.includes(`gatsby`)
-            ),
-          }
-          createNodeField({
-            node,
-            name: `starterShowcase`,
-            value: starterShowcaseFields,
-          })
-        })
-        .catch(err => {
-          reporter.panicOnBuild(
-            `Error getting repo data for starter "${repoStub}":\n
-            ${err.message}`
-          )
-        })
-    }
-  } else if (node.internal.type === `CreatorsYaml`) {
-    // Creator pages
-    const validTypes = {
-      individual: `people`,
-      agency: `agencies`,
-      company: `companies`,
-    }
-
-    if (!validTypes[node.type]) {
-      throw new Error(
-        `Creators must have a type of “individual”, “agency”, or “company”, but invalid type “${node.type}” was provided for ${node.name}.`
-      )
-    }
-    slug = `/creators/${validTypes[node.type]}/${slugify(node.name, {
-      lower: true,
-    })}`
-    createNodeField({ node, name: `slug`, value: slug })
-  }
-  // end Creator pages
-  return null
+exports.onCreateNode = helpers => {
+  sections.forEach(section => section.onCreateNode(helpers))
 }
 
 exports.onPostBuild = () => {
